+++ src/main/java/br/com/techchallenge/adapters/converter/RestauranteDTOConverter.java
@@ -1,6 +1,6 @@
 package br.com.techchallenge.adapters.converter;
 
-import br.com.techchallenge.adapters.dto.RestauranteRequestDTO;
+import br.com.techchallenge.adapters.dto.Restaurante.RestauranteRequestDTO;
 import br.com.techchallenge.infra.entity.RestauranteEntity;
 import org.springframework.stereotype.Component;
