+++ src/main/java/br/com/techchallenge/adapters/dto/Restaurante/RestauranteResponseDTO.java
@@ -1,4 +1,4 @@
-package br.com.techchallenge.adapters.dto;
+package br.com.techchallenge.adapters.dto.Restaurante;
 
 public record RestauranteResponseDTO(
         String nome,
