+++ src/main/java/br/com/techchallenge/adapters/dto/donoRestaurante/DonoRestauranteResponseDTO.java
@@ -1,4 +1,4 @@
-package br.com.techchallenge.adapters.dto;
+package br.com.techchallenge.adapters.dto.donoRestaurante;
 
 public record DonoRestauranteResponseDTO(
         String nome,
